+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 * Change to search default editors when command and env variables checks fail
 * Change to rename :show_menu option as :hide_menu to indicate default behaviour
+* Change default editors list to be sorted in alphabetical order
 
 ### Fixed
 * Fix Editor#exist? to correctly check command with an absolute path